@@ -1,18 +1,18 @@
 name: dio_smart_retry
 description: Retry library for Dio and Dio package made with love. By default, the request will be retried only for appropriate retryable http statuses.
-version: 5.0.0
+version: 5.3.3
 repository: https://github.com/rodion-m/dio_smart_retry
 issue_tracker: https://github.com/rodion-m/dio_smart_retry/issues
 homepage: https://github.com/rodion-m/dio_smart_retry
 documentation: https://github.com/rodion-m/dio_smart_retry#contents
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
-  dio: ^5.0.0
+  dio: ^5.3.3
   http_parser: ^4.0.2
-  path: ^1.8.2
+  path: ^1.8.3
 
 dev_dependencies:
   dart_code_metrics: ^5.6.0
